backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
aiofiles
biopython
numpy
scipy
pydantic